Fix binary compatibility
Change type of Cookie.maxAge back to int to maintain binary
compatibility for drop in fixes.

+++ src/main/java/org/jboss/netty/handler/codec/http/DefaultCookie.java
@@ -154,7 +154,7 @@ public class DefaultCookie extends org.jboss.netty.handler.codec.http.cookie.Def
     }
 
     @Deprecated
-    public long getMaxAge() {
+    public int getMaxAge() {
         return maxAge();
     }
